Allow influencing the amount of snapshots in a replay file
This is mostly implemented to allow a very low number of snapshots, for
cases where this is not important: most particularly, an auto-saved
replay.

For some reason, for a low amount of snapshots, the amount is sometimes
larger than what I expected, but guess this was already the case with
the fixed number of 10 from before this patch, as the only thing that
this patch does is make that fixed number variable.