Use shared_from_this in MSXWatchIODevice
Create a local shared_ptr to the 'this' object (to keep it alive in case the
watchpoint callback removes the current watchpoint) using 'shared_from_this'
instead of making a temporary copy of all watchpoints.

This change was already suggested in the code comments. Though at that time we
weren't using c++11 yet and our own implementation of shared_ptr didn't support
shared_from_this.