[1/10] Replace 'double' with 'float' in OSD code
In the OSD-widget related code replace all (64-bit) double types with (32-bit)
float types. 'float' should be more than accurate enough for what we need. (In
fact openGL uses 32-bit for all coordinate calculations, and there 32-bit is
never a problem, so it certainly won't be a problem for simple 2D
calculations).

On x86 hardware there is not that much advantage in using float instead of
double: a bit less memory usage, maybe slightly more opportunities for SIMD
vectorization, but the actual floating point calculations are about as fast for
double as for float. However on CPUs without hardware floating point support
this can make a difference.

The first patches in this series (replace double->float) also make
preparartions for the later patches (replace scalar->vector).