[7/10] Use vector instead of scalar calculations in console/
Let's summarize the changes in this patch with an example:

before:
    void transform(float x, float y, float& outX, float& outY) {
        outX = x * scale + offsetX;
        outY = y * scale + offsetY;
    }
after:
    vec2 transform(vec2 position) {
        return position * scale + offset;
    }

So we replace scalar calculations with vector calculations. Notice how in the
first example the calculation for the x component is almost copy/pasted for the
y component. That redundancy is eliminated in the second example.

In the first example we needed to return two scalar values, we used output
parameters for this. In the second example there's only a single vector to
return, so we can use the more natural return syntax. This might enable better
optimizations because in the first example the compiler might not be able to
prove the output parameters don't alias any of the other variables.

This patch makes changes in the console/ subdirectory. Later patches will
tackle other parts of the code.