Optimize V9990 palette init on soft-float targets
Copied the palette-calculation optimization for the V99x8 code to V9990. See
revision 3a0483c30 for more details on the V99x8 part.

To apply gamma correction we need to call the pow() function for each of the 3
RGB color components for each of the 32768 colors in the V9990 palette. On
targets without hardware floating-point support these calculations take quite a
bit of time (luckily it's only required once during initialization). However in
case the color_matrix setting is the identity matrix (the default value), the
transformation for the 3 RGB components is identical (and independent of the
other components) and we can pre-calculate the gamma correction for the 32
intensity values (instead of 3 x 32768 values).

As already mentioned, the exact same optimization was already done for V99x8.
It matters mostly for small devices without hardware floating point support.
Maybe V9990 emulation is not used much on such devices. But still this patch
makes the code more similar between V99x8 and V9990.