Fixed bug when calling freeAligned() on nullptr
We have 3 implementations for mallocAligned()/freeAligned():
- when available we use posix_memalign()/free()
- with visual studio we use _aligned_malloc()/_aligned_free()
- and as a fallback we use malloc()/free() + extra book keeping for the alignment.
(Only?) when compiling openMSX with mingw (gcc on windows) the 3rd version is
used.

In the first two variants freeAligned(nullptr) is allowed (and does nothing).
Though in the 3rd variant (in our manual book keeping code) it triggered an
assert.

I've extended the 3rd variant to allow freeing nullptr's.