Replace our Semaphore wrapper with std::mutex
We didn't actually need the counting-feature of a Semaphore. The only values
our semaphores took were 0 and 1, and in that case a mutex is fine as well.
(Maybe in the past we had other uses as well? I didn't check).

C++11 offers std::mutex (actually our Semaphore class was already rewritten to
use that). So this patch uses std::mutex directly. Most of this patch are
mechanical replacements:
  Semaphore  -> std::mutex
  ScopedLock -> std::lock_guard<std::mutex>  or
                std::unique_lock<std::mutex>

The difference between lock_guard and unique_lock is small:
  https://stackoverflow.com/questions/20516773/stdunique-lockstdmutex-or-stdlock-guardstdmutex
Short summary:
  * unique_lock is a strict superset of lock_guard
  * the more limitted functionality of lock_guard may more clearly convey the
    intend of the code.
So I'm using lock_guard were possible and lock_guard were needed (only in
EventDistributor).