Replace volatile bool with atomic<bool>
In the CPUCore code we used a 'volatile bool' to communicate a flag from one
thread to another. This patch replaces that with an 'atomic<bool>'.

Although the 'volatile bool' approach seemed to work in practice, the c++
standard does not guarantee it will. To communicate between threads atomic are
needed (they make guarantees about memory orderings). See here for some
background info:
  https://stackoverflow.com/questions/8819095/concurrency-atomic-and-volatile-in-c11-memory-model
Though the old code was written well before atomics became available in c++11.