Remove back-pointers from inner to outer objects
Some time ago we had code patterns like this:
    class Foo {
        ...
        class Bar : public Schedulable {
            virtual void executeUntil() {
                foo.bla();
            }
            ...
            Foo& foo; // back-pointer to Foo
        };
        std::unique_ptr<Bar> bar; // owning pointer to separate object
        ...
        void bla() { ... }
    };

These were transformed to:
    class Foo {
        ...
        class Bar : public Schedulable {
            virtual void executeUntil() {
                foo.bla();
            }
            ...
            Foo& foo; // back-pointer to Foo
        } bar; // bar object directly contained inside Foo
        ...
        void bla() { ... }
    };

So initially there were two separately allocated objects that had pointers to
one another (one owning pointer and one back-pointer (actually a reference)).
Later the second object became part of the first, so there's only one
allocation anymore and the owning pointer is removed, however the back-pointer
remained.

This patch transforms the above pattern further to also get rid of the
back-pointer.
    class Foo {
        ...
        class Bar : public Schedulable {
            virtual void executeUntil() {
                auto& foo = OUTER(Foo, bar); // <-- calculate back-pointer
                foo.bla();
            }
            ...
            // no need to store back-pointer anymore
        } bar
        ...
        void bla() { ... }
    };

The back-pointer is still needed, but it can be calculated instead of storing
it as a member variable. The 'bar' object lives at a fixed offset inside the
'Foo' object, so we can calculate the Foo-pointer from the Bar-pointer by
subtracting that offset.

See comments in utils/outer.hh for more details.