Use -std=c++11 when compiling with GCC
We used -std=gnu++0x before, which enabled C++11 features prior to the
C++11 spec being finalized. Since GCC 4.7 is now required, we can use
-stc=c++11 instead.

Note that this change also disables support for GNU extensions, but
that should not be a problem and for portability it is better to avoid
compiler extensions as much as possible anyway.