Don't mark this-pointer restrict in OutputBuffer
Remove the 'restrict this-pointer' annotations from the OutputBuffer class.
This fixes compilation errors in clang. The regression was likely introduced by
revision 79aeb0d (but I didn't verify).

Gcc (and vs) seem to behave slightly different compared to clang when marking
the 'this' pointer restrict. Maybe not so surprising because this is not part
of the c++ standard. Strictly speaking the whole restrict feature is not
covered by the c++(-11) standard, but it is a widely supported extension (by
gcc, visual-studio, clang and probably many other c++ compilers).

Removing these annotations only seems to have a very small negative impact on
the quality of the generated code (I verified using gcc-4.9). In the past I
remember seeing a bigger effect. Possibly the alias analysis passes in gcc have
gotten more powerful since that time. Anyway because of this I didn't bother
trying to keep these annotations for gcc and visual-studio.