Use -Wno-invalid-offsetof instead of MY_OFFSETOF
Revision 49f27fea656 introduced a MY_OFFSETOF macro. It works exactly the same
as the standard offsetof macro, except that it was written in such a way to
suppress gcc/clang compiler warnings (because we're using it in a not 100%
standard compliant way).

This patch removes MY_OFFSETOF (replaces it with offsetof) and instead
suppresses the warning using the -Wno-invalid-offsetof compiler flag. I checked
that
  gcc-4.7, gcc-4.8, gcc-4.9, gcc-5
  clang-3.4, clang-3.5, clang-3.6
all support this flag and that they compile openMSX warning-free.

The description of the flag in the gcc manual is interesting:
  -Wno-invalid-offsetof (C++ and Objective-C++ only)
     Suppress warnings from applying the "offsetof" macro to a non-POD type.
     According to the 2014 ISO C++ standard, applying "offsetof" to a non-standard-
     layout type is undefined.  In existing C++ implementations, however, "offsetof"
     typically gives meaningful results.  This flag is for users who are aware that
     they are writing nonportable code and who have deliberately chosen to ignore the
     warning about it.

     The restrictions on "offsetof" may be relaxed in a future version of the C++
     standard.
(At least to me) this description reassures that using offsetof in this not
strictly standard compliant way is unlikely to cause problems in practice.

Thanks to GuyveR800 for the suggestion.