Fixed potential alignment bug
Visual studio generated warnings like this:
    C4316: 'openmsx::IDEHD' : object allocated on the heap may not be aligned 16
The problem is that IDEHD contains a member variable of type 'AlignedBuffer',
we've (manually) increased the alignment of that type to 16 bytes. Though on
some systems heap objects are only 8 bytes aligned, resulting in a possible
misalignment of the AlignedBuffer variable.

The warning is correct, this is a bug. Though I believe it didn't cause
problems in practice for the following two reasons:
 - On 64-bit systems (I verified both Linux and Windows) malloc returns 16-byte
   aligned buffers. So there's no problem.
 - On 32-bit systems, malloc only returns 8-byte aligned data. Though (on x86)
   this misalignment only causes problems when SSE instructions are used, and
   by default the compiler does not generate those (on gcc you'd have to pass
   a -march or -msse2 flag, in 32-bit mode these are not enabled by default).

This patch fixes the problem by only giving AlignedBuffer 'max_align_t'
alignment. This is the alignment guaranteed by malloc().  (I hope visual studio
accepts the current syntax in the patch, I could only test that gcc/clang do).

Background info:
The 'AlignedBuffer' object inside IDEHD (and other disk related classes)
represents a sector, so a buffer of 512 bytes. One possible fix is to just
keep the default alignment (1 byte instead of 8 or 16). I prefer to keep the
stricter alignment for two reasons:
 - Sector buffers are sometimes interpreted as directory entries or other FAT
   filesystem structures. Those structures contain 16- or 32-bit values. On
   some targets (e.g. ARM) reading such values is more efficient when they are
   properly aligned (and if you're not careful reading misaligned 16/32 bit
   values on (older) ARM processors causes an abort).
 - We have many places where we memcpy() sector buffers. If the compiler knows
   about the alignment of the source and destination buffers it can generate
   more efficient (and smaller) code. (Though recent gcc versions have an
   optimization regression, see https://gcc.gnu.org/bugzilla/show_bug.cgi?id=50417)