Fixed(?) visual studio compiler error
As I mentioned in the previous commit message, I couldn't test that patch on
windows. And indeed visual studio doesn't like it. It seems that
__declspec(align(#)) only accepts integer constants for '#', and not c++
constant expressions (as the equivalent constructs in gcc/clang do).

This patch adds a hack for visual stdio. That can hopefully be removed again
once we switch to visual studio 2015. That one should (finally) support the
c++11 alignas/aligof features.