Don't store prompt in console history
The console command history is saved on disk (in the file
~/.openMSX/persistent/console/history.txt). While openMSX is running the same
data is also present in memory. Before this patch, in the in-memory version,
all commands were prepended with the command prompt (usually '> '). Now that
prompt is removed.

This change make some functions simpler, others more complex. Combined it's
still a slight simplification.