[3/n] Read user and system softwaredb.xml in a single buffer
In a later patch in this series we will refer to strings using indices in a
buffer. For this to work we need both user and system softwaredb.xml strings to
reside in a single buffer. (And after the parsing has finished, we don't
know/care anymore whether a RomInfo originated from the user or the system
softwaredb.xml).