Compiled with -Wsuggest-override and fixed all warnings
I did not (yet) make -Wsuggest-override part of the default set of warnings
because gcc-4.7 does not yet support this flag.

Most changes in this patch are obvious, only one type of change might need more
explanation:

In some places we have an inheritance hierarchy with (at least) 3 levels. The
top level base class declares a (pure) virtual method e.g.
    virtual void myMethod() = 0;
The middle class does not yet implement that method, but the bottom subclass
does. In some places, we repeated that (fully redundant) declaration in the
middle class. This repetition emphasizes the point that this class has that
particular method but does not yet implement it (makes it easier to understand
the middle class without having to look at the top class).

The -Wsuggest-override option warns about such constructs. In some places, when
the middle class on itself is important enough, I've chosen to keep the
(redundant) repetition but add the 'override' keyword:
    void myMethod override = 0;
In other places, e.g. when the middle class is more an implementation detail,
I've chosen to remove the repetition. I'm not sure yet which style I like best.