Enable -Wold-style-cast
Code base was already warning-free when enabling this flag. At least the
linux-side, I suspect there may be some issues left in the windows-specific
code, but we should fix those.

This is a follow-up on ba0f6c438a05926 (May 2015). At the time we couldn't
enable -Wold-style-cast yet because it didn't work well with gcc-4.7. Support
for that compiler has been dropped, so now we can enable this warning.