[4/4] use String32 instead of string_ref in RomInfo
The previous patches in this series already mentioned the key ideas, but I'll
repeat/summarize them again below.

When parsing the (user+system) softwaredb.xml files we roughly take the
following approach:
1) Read the full XML file in a memory buffer.
2) Parse the file using our rapidsax parser, this calls various callbacks.
3) These callbacks construct an additional data structure: a vector of RomInfo
   objects. Those objects refer to the strings in the XML buffer. (This means
   the buffer allocated in step 1 cannot be freed once parsing is done).

On a 64-bit system the size of a string_ref object is 16 bytes. Each RomInfo
objects has 6 such string_ref objects (plus some other stuff). The strings that
are referred to are typically not very large. So this means that the meta-data
(I mean the vector of RomInfo objects) approaches the size of the original XML
file. I measured about 600kB for the raw XML data and 430kB for the meta-data.

In this patch series we greatly reduce the size of the meta-data. To do this we
combine two ideas:
- Instead of storing the size of the strings (in string_ref) we zero-terminate
  the strings (so we modify the in-memory copy of the XML file). This halves
  the size of the string_ref objects. (Operations on zero-terminated strings
  are often slightly slower compared to sized-strings, but these strings are
  only infrequently used so this doesn't matter.)
- On 64-bit systems a pointer is 8-bytes. But 4 bytes are already sufficient to
  address 4GB of memory (I don't expect the size of the softwaredb.xml file to
  grow over 1MB, that's far below 4GB). So if we replace pointers with indices
  in a buffer we can again halve the cost of each string_ref object (but only
  on 64-bit systems).

After these changes the size of the meta-data shrinks to about 176kB (from
430kB). That's not exactly 4x less because there are also some other members in
RomInfo that remain unchanged.

I also measured the startup speed using 'perf' and 'callgrind'. Both methods
indicate small (less than 0.5%) but repeatable speedups. Speed was not the goal
of this patch, but it's a nice bonus.