[2/4] Tweak rapidsax XML-parser
One of the properties of our rapidsax parser is that it doesn't make any string
copies. Instead it does XML-entity replacement in-place (in the original source
buffer). The parser then passes references to parsed pieces of text (text or
attributes) as string_ref objects to various callback methods.

Before this patch series, the RomDatabase code stored those string_ref's inside
RomInfo objects. A string_ref is basically a pointer and a size, it tells where
to find the string and how long it is. In this patch series we get rid of the
size-part. Instead of storing the size of each string we zero-terminate them.
(We store many RomInfo's and typically use only very few, so this is a good
trade-off).

Because of the structure of XML files there's always room in the buffer to
store the 0-terminator character. Though we can only overwrite the content of
the buffer after the original character has been processed by the parser. This
patch changes the order in the rapidsax code a bit so that it's always OK for a
callback to 0-terminate the text it just received.