Tweak '"::" + machineID + "::" + name' string calculation
Replace
    "::" + machineID + "::" + name
with
    machineID + name
but store the subexpression
    "::" + machineID + "::"
inside the (new) machineID variable.

This patch has 1 unintentional side effect. (In related code) before we called
    namespace create ::foo {}
now that becomes
    namespace create ::foo:: {}
and similar for "namespace delete", but I verified that both variants have the
same effect.

The motivation for this patch is the following: when looking at some profile
results I noticed that the Tcl setting related functions perform relatively
many string manipulations. This patch is one small step to improve that. (I'm
still playing with other improvements but they are not quite ready yet).