[1/15] Remove CommandController::findSetting()
We want to get rid of the findSetting() methods in the CommandController,
MSXCommandController and GlobalCommandController classes. This initial step
removes it from the base class (CommandController) and makes the findSetting()
methods in the subclasses non-virtual.

It turns out there was only one location (inside UserSettings) that used the
virtual-nature of findSetting(), and it was easy to remove it.

--

Because this is the first patch of this series, I'll also give a _very_ brief
overview:

  [1/15] Remove CommandController::findSetting()
  [2/15] Added GlobalCommandController::getSettingsManager() helper function
  [3/15] Added SettingsManager::findSetting(prefix, baseName)
  [4/15] Don't use MSXCommandController::findSetting() in ProxySetting
  [5/15] Don't use MSXCommandController::findSetting() in transferSettings()
  [6/15] Remove MSXCommandController::findSetting()
  [7/15] Replace 'hash_set settingMap' with 'vector settings'
  [8/15] Also remove GlobalCommandController::findSetting()

These are mostly preparation patches. After these MSXCommandController plays
a less critical role in handling of machine specific settings. Especially in
the naming of the settings.

  [9/15] Made Setting class aware of fullName/baseName
  [10/15] No longer need to pass name
  [11/15] Change internal representation of Setting::xxxName
  [12/15] Use Tcl_Objxxx() functions / pass TclObject parameters
  [13/15] Replace hash_map with hash_set

These patches form the core of this series. They change the interaction with
the Tcl library (more time and space efficient). It also removes some
redundancy from some of our datastructures.

  [14/15] Enhance tab-completion for setting names
  [15/15] Removed changeSetting() methods from XXXCommandController

These are mostly cleanups made possible by the previous patches.

--

One of the goals of this patch series was to improve performance. So I made
some measurements. I measured startup time + boot to BASIC:
- There's no measurable change in execution time. This is expected because this
  scenario doesn't trigger the Setting code much. I guess very few scenario's
  do, but who knows what people do in Tcl scripts. (Actually I did measure a
  small improvement, but it's close to the measurement noise, so likely not
  significant).
- When looking at memory usage there is a measurable improvement: there are
  about 400 less malloc() calls. For a reduction in memory usage of about 35kB.
  This can be explained by the more compact data structures and the more
  efficient string-sharing between openMSX and Tcl.
- The size of the openMSX executable has reduced slightly (4kB). Though this is
  not significant.
- The last item is not objectively measurable, but IMHO the new code structure
  is also easier to understand because the difference between global and
  machine specific settings has become smaller. And code paths that previously
  had to take a detour via MSXCommandController now go directly to the
  appropriate classes.