Fixed initialization of String32 on 32-bit platforms
In the RomDatabase parsing code we initialized String32 objects like this:
    title = 0;
This compiles fine (on both 32 and 64 bit platforms). It also happens to do the
correct thing on 64 bit systems. However on 32 bit systems when a romdb entry
with empty title, year, country or company field is printed (depending on how
exactly it's printed) it may crash. So it may only rarely trigger, but a bug is
a bug.

Fix is easy: always initialize these objects via the toString32() helper
function.