Fake just enough of the Mitsubishi ML-TS2 to get it to boot
There seems to be 4 registers located at addresses 0x7f00-0x7f03. The z80 code
in the rom both reads and writes these registers (I only looked at a very small
fraction of the code).

Before this patch the boot code was in an endless loop reading location 0x7f01
and checking bit 3. Now, after this patch, that bit reads as '1' and that's
enough to get past this loop into BASIC.