[1/2] Use 'small_compare' to speedup RomDatabase parsing
Added a small_compare() utility function and use it to speedup parsing of the
rom database. I measured about 1% faster total startup time.

While parsing the rom database we often need to compare a string against a
small fixed string. E.g. compare a tag-name against a set of small known
strings like "software" or "country". Before this patch we used the generic
string comparison function (which internally uses memcmp()) and profiling
showed that a lot of time was spend in the __mem_cmp_sse4_1 function (on my
system). This function is very well optimized, but it is a generic memory
comparison function. In our case both the length and the content of the string
are known at compile-time. So maybe we can do better than the generic memcmp()
function.

After this patch we use small_compare() instead of memcmp(). Look at the
comments in small-compare.hh in this patch for a detailed description. But a
short summary (ignoring some details): to match the string "year" we do a
4-byte load and compare that against the constant 0x72616579 (the ascii
characters for 'y','e','a','r' translated in hex and concatenated). This
requires only 2 instructions and is much faster than a call to memcmp().