[1/25] Replace set with vector (only has few elements)
The MSXMotherboard class has a set<string> for the 'usernames' of the different
hardware extensions. This username stuff is much less used than originally
intended (10 years ago), but it's responsible for giving unique 'persistent'
subdirectories in case the same extension is inserted multiple times in the
same machine. E.g. a machine with two fm-pac extensions: one will store the
SRAM content in ~/.openMSX/persistent/untitled1/fmpac.pac, the other in
~/.openMSX/persistent/untitled2/fmpac.pac.

These 'usernames' (untitled1, untitled2, ...) have to be unique, so originally
we stored them in a std::set. Though because typically there's only one item in
this set (only rarely 2, and practically never more than 4) it's more efficient
to use a std::vector.

--

Because this is the 1st patch in this series I'd also like to give a small
overview:

   1 Replace set with vector (only has few elements)
   2 Remove Interpreter::commandTokenMap
   3 Removed name member from ProxyCommand

These are things that I discovered while working on this patch series, but are
fairly independent from the core of this series (see next paragraphs), so I
moved these patches to the front.

   4 Added EqualTo functor
   5 Added hash-set.hh and hash_map.hh
   6 Add XXHasher functor

These patches add new functionality.

   7 StringMap->hash_map for MSXMotherBoard::sharedStuffMap
   8 StringMap->hash_map for MSXMotherBoard::userNames
   9 StringMap->hash_set for OSDWidget::subWidgetsMap
  10 StringMap->hash_map for Debugger::debuggables
  11 StringMap->hash_set for Debugger::probes
  12 StringMap->hash_map for romTypeMap
  13 StringMap->hash_map for unknownTypes
  14 StringMap->hash_map for CommandLineParser::itemMap
  15 StringMap->hash_set for InfoCommand::infoTopics
  16 StringMap->hash_map for {Global,MSX}CliComm::prevValues
  17 StringMap->hash_set for decompressCache
  18 StringMap->hash_map for loaderMap/initializerMap
  19 StringMap->hash-set for MSXCommandController::settingMap
  20 StringMap->hash_set for MSXCommandController::commandMap
  21 StringMap->hash_map for GlobalCommandController::commandCompleters
  22 StringMap->hash_map for GlobalCommandController::{commands,proxyCommandMap}
  23 StringMap->hash_map for SettingsManager::settingsMap

The bulk of this patch series gradually replaces StringMap with the newly added
hash_set or hash_map.

  24 No longer need StringMap in KeyRange
  25 Removed StringMap

In the end StringMap is no longer used and it can be removed.