[25/25] Removed StringMap
StringMap is no longer used, remove it.

StringMap was a hash-map specialized for the case that the key was a string: it
would use a single memory allocation to store both the mapped-value and the
(variable sized) string content.

The previous patches have replaced all uses of StringMap<x> with either
hash_set<x> or hash_map<string, x>. In case of:
- hash_set<x>: The string was part of the mapped-value, so storing a copy (even
  if done efficiently) is not very useful.
- hash_map<string, x>: Here the old StringMap<x> may have been more efficient,
  but having two hash-maps is probably not worth the extra complexity. Also,
  starting from gcc-5, std::string performs the small-string-optimization and
  that greatly lessens the StringMap advantage.