[3/25] Removed name member from ProxyCommand
Revision d5c3ab1d added a single ProxyCommand object that worked for all
machine specific commands (delegates from the global to the machine specific
command). In rev 7a0a25b9 that was changed (because of issues with renaming
command) and a name member was added.

By refactoring (= also simplifying) the command registration stuff it's possible
to move that extra name member back to the base class.