[2/25] Remove Interpreter::commandTokenMap
The main theme in this patch series is to replace StringMap with hash_map or
hash_set (both are introduced later). Though in this case it was possible to
remove commandTokenMap altogether:

Remove the commandTokenMap member from Interpreter, instead store (a single)
token in each Command object. (This means that each Command object can at most
be registered once, but that's ok).