Speedup looking-up OSD widgets by name
In revision de8caffa (Aug 2009) we speedup lookup by name of OSD widgets.
Because in some 'extreme' cases it was a bottleneck (only happens with
'extreme' Tcl scripts that change OSD properties very often; simply drawing
doesn't require lookup by name). This patch speeds it up a little further.

A OSDWidget contains a (owning) vector of child OSDWidget objects, those are
sorted on increasing z-coordinate (for fast drawing). Revision de8caffa also
added a (non-owning) (hash)map to speedup lookup by name. These maps exist per
widget, so at each hierarchical level you can do a O(1) lookup. Thus a
full hierarchical name lookup is O(hierarchy-depth).

It turns out we only need lookups on the full hierarchical name (IOW we never
start a lookup in a subtree). So we can replace all those maps with a single
(hash)map in the top element. This makes all name lookups O(1).

--

After the above change a widget no longer needs to know it's own name, so this
patch also removes the name member from the OSDWidget class.

The name is now only stored in the top lookup map. Note that it's the _full_
name, so the common prefix of (long) hierarchical names is now duplicated in
memory.

To compensate for this possible higher memory usage the names are now stored as
TclObjects instead of std::strings. Often the names of the widgets are in some
form present in Tcl scripts (e.g. as the value of some variable). By keeping
the name as a TclObject we avoid any string conversions (an additional speedup)
and also (often) avoid the extra memory cost for the name (because TclObjects
are reference counted).