Added operator==(TclObject, string_ref)
Before when comparing a TclObject with a string(_ref) you had to write:
   if (myTclObject.getString() == "someValue") { ...
After this patch that can be simplified to:
   if (myTclObject == "someValue") { ...

This allows to simplify the code in many places. Though the main motivation for
this change is that it allows to have a hash_map with TclObject keys and use
string_ref (instead of TclObject) to lookup elements (an alternative was to use
a custom equality comparison functor instead of the default one). This will be
used in one of the later patches.