Don't crash when calling xxhash(nullptr, 0)
This calculates the hash-value of a 0-sized memory block. So whatever pointer
you pass (even a nullptr) should be fine.

The problem was the following. Internally the function did:
    uint8_t* end = p + size;
    while (p <= (end - 4)) {
       ... process 4 bytes at-a-time
    }
    ... process the remaing bytes
So as long as we're 4 bytes before the end we can process a chunk of 4 bytes
(instead of byte-per-byte). Though when p=nullptr and size=0 then end=nullptr,
end-4 wraps around and the comparison goes wrong.

To solve this I changed the comparison to
    while ((p + 4) <= end) { ...
This goes wrong when the object is within 4 bytes from the upper limit of the
address-space, but I don't think this can happen on any of our supported
platforms.