Fix Trackball movement in joytest-v2.3b
Technically there wasn't anything wrong in the trackball emulation. However
because of the discretization of motion the heuristics used by the MSX-BIOS (to
distinguish a mouse from a trackball device) went wrong. See comments in the
code for more details.

To fix this we changed the way how host mouse-movement is translated into
emulated trackball-movement: we smear out the host movement slightly over
emulated time, so that when looking at small emulated time intervals, the
trackball never moves too fast.