IRQHelper cleanups
1) Added a helper method 'void IRQHelper::set(bool)' which, depending on the
argument, calls either set() or reset(). So we can replace the pattern
    if (myCondition) {
        myIRQ.set();
    } else {
        myIRQ.reset();
    }
with
    myIRQ.set(myCondition);

2) In some places we did
    if (!myIRQ.getState()) {
        myIRQ.set();
    }
But internally the set() method already does this same check, so the above
pattern can be simplified to
    myIRQ.set()

3) And similar for
    if (myIRQ.getState()) {
        myIRQ.reset();
    }
->
    myIRQ.reset();

This is a pure cleanup patch, there should be no functional changes.

I noticed these recurring patterns while working on the new YM2148 code. But
since this cleanup is independent of that work I'm committing it separately.