[9/15] Made Setting class aware of fullName/baseName
Before this patch each Setting object had a single name. After this patch a
Setting has both a fullName and a baseName.

For global settings nothing changes: the fullName and the baseName are exactly
the same (and the same as before this patch). However for machine specific
settings there is a difference: the fullName is the fully qualified name
including the machine prefix and baseName is the last part of the name (same as
before this patch). For example:
  fullName -> ::machine1::PSG_volume
  baseName -> PSG_volume

On itself this patch is not very useful, but it's a preparation for later
patches in this series. It will allow to remove some redundancy from some
datastructures and trigger less string conversions while manipulating Tcl
settings.

baseName is always a strict suffix of fullName, so it is possible to use a more
compact representation than two std::string variables. Though a later patch in
this series will change the representation to two TclObjects.