Work around (hack) crash in MidiInReader
Exiting openMSX while the MidiInReader pluggable is still active often results
in a crash. This was because the helper thread (for reading the file) is still
using the already deleted MidiInReader object.

This patch is more a hack than a proper solution: it will work as long as only
one MidiInReader is active at the same time. But in principle it's possible to
start multiple MSX machines (in two tabs) and in each use a different
MidiInReader.

A better solution would be to get rid of the helper thread. (Maarten was
working on exactly this, but I don't know the status of that work).