[12/15] Use Tcl_Objxxx() functions / pass TclObject parameters
Change various functions to take 'TclObject' instead of 'std::string'
parameters. And in the implementation of (some of) those functions call the
object instead of the C-string variant of the Tcl library functions.

The Tcl library offers several variants for (some of) its functions. For
example there is Tcl_SetVar2Ex() and Tcl_ObjSetVar2(). The main difference
between these variants is that the former takes a 'const char*' type as name
parameter while the latter takes a 'Tcl_Obj*' type.

When peeking in the internal implementation of the Tcl_SetVar2Ex() function, we
see that it first creates a Tcl_Obj out of the C-string and then calls the
Tcl_ObjSetVar2() function. Creating a Tcl_Obj involves a strlen(), malloc() and
memcpy() operation. And later a free() when the temporary Tcl_Obj is destroyed
again. So if we directly call the Tcl_ObjSetVar2 version ourselves we avoid
some overhead.

When peeking in the internal Tcl datastructures that store the Tcl variables
(corresponds to our settings) we see that Tcl_Obj's are used to store the name
of the variables. When we pass a Tcl_Obj ourselves to Tcl_ObjSetVar2 (instead
of letting one create internally via Tcl_SetVar2Ex) the storage for that name
can be shared between the internal Tcl datastructure and the Setting::fullName
member variable (because Tcl_Obj's are internally reference counted). So next
to a speed advantage there's also a storage advantage.