Change set/map in HotKey to vector
Insert/lookup/erase in a set/map is O(log(n)), while the same operations in a
vector are O(n). Though this theoretical time-complexity is for 'large' values
of 'n'. In practice for n < 100 the operations on a (unsorted) vector are
actually faster than on set/map. Vector also uses less memory (part of the
reason why it's faster).

So in general small sets/maps are better implemented as a vector. This patch
applies this principle to the HotKey class. HotKey contains 2 maps ('cmdMap',
'defaultMap') and 2 sets ('boundKeys' and 'unboundKeys'). Typically these all
contain less than 10 elements.

When using a few helper functions, set/map operations on a vector are not less
readable than on a set/map.